Add error handling, config validation, and improved path safety

+++ jekyll-rmagick.rb
@@ -26,6 +26,7 @@
 
 require 'rmagick'
 require 'fileutils'
+require 'pathname'
 
 include Magick
 
@@ -44,14 +45,10 @@ module Jekyll
 
         class Generator < Jekyll::Generator
 
-            $pathformat = "%{filename}-%{size}.%{extension}"
+            PATH_FORMAT = "%{filename}-%{size}.%{extension}"
 
             def symbolize_keys(hash)
-                result = {}
-                hash.each_key do |key|
-                    result[key.to_sym] = hash[key]
-                end
-                result
+                hash.transform_keys(&:to_sym)
             end
 
             def format_output_path(dest, image_path, size)
@@ -71,37 +68,57 @@ module Jekyll
 
             def generate(site)
 
-                dest = site.config['destination'] + "/assets/"
-                src = site.config['source'] + "/" + site.config['jekyll-rmagick']['source_dir'] + "/"
+                jekyll_rmagick_config = site.config['jekyll-rmagick']
+                return unless jekyll_rmagick_config
 
-                prefix = site.config['jekyll-rmagick']['prefix']
-                sizes = site.config['jekyll-rmagick']['spec']
-                FileUtils::mkdir_p dest # directory needs to exist or IM won't put images there.
+                source_dir = jekyll_rmagick_config['source_dir']
+                return unless source_dir
+
+                prefix = jekyll_rmagick_config['prefix'] || 'image'
+                sizes = jekyll_rmagick_config['spec']
+                return unless sizes && sizes.is_a?(Array)
+
+                dest = File.join(site.config['destination'], 'assets')
+                src = File.join(site.config['source'], source_dir)
+                FileUtils.mkdir_p dest
 
                 site.posts.each do |post|
                     if post.data['img_src']
 
-                        source_file = src + post.data['img_src']
-                        original = ImageList.new(source_file)
-                        original.strip!
+                        source_file = File.join(src, post.data['img_src'])
+                        unless File.exist?(source_file)
+                            site.logger.warn "jekyll-rmagick: Source file #{source_file} does not exist"
+                            next
+                        end
 
-                        sizes.each do |spec|
+                        begin
+                            original = ImageList.new(source_file)
+                            original.strip!
 
-                            filename = format_output_path($pathformat, source_file, spec['meta'])
-                            output_file = dest + filename
-                            rel_path = '/assets/' + filename
-                            image = original.copy
-                            image.resize!(spec['width'], spec['height'])
+                            sizes.each do |spec|
+                                next unless spec['meta'] && spec['width'] && spec['height']
 
-                            if site.config['jekyll-rmagick']['brightness-mod']
-                                image.modulate!(site.config['jekyll-rmagick']['brightness-mod'])
+                                filename = format_output_path(PATH_FORMAT, source_file, spec['meta'])
+                                output_file = File.join(dest, filename)
+                                rel_path = File.join('/assets', filename)
+                                image = original.copy
+                                image.resize!(spec['width'], spec['height'])
+
+                                if jekyll_rmagick_config['brightness-mod'] && jekyll_rmagick_config['brightness-mod'].is_a?(Numeric) && (0.0..1.0).include?(jekyll_rmagick_config['brightness-mod'])
+                                    image.modulate!(jekyll_rmagick_config['brightness-mod'])
+                                end
+
+                                quality = jekyll_rmagick_config['quality'] || 75
+                                image.write(output_file) { self.quality = quality }
+                                image.destroy!
+                                site.static_files << ImageFile.new(site, site.source, '/assets', filename)
+                                post.data[prefix + "-" + spec['meta']] = rel_path
+
                             end
 
-                            image.write(output_file) { self.quality = site.config['jekyll-rmagick']['quality'] || 75 }
-                            image.destroy!
-                            site.static_files << ImageFile.new(site, site.source, '/assets', filename)
-                            post.data[prefix + "-" + spec['meta']] = '/assets/' + filename
-
+                        rescue => e
+                            site.logger.error "jekyll-rmagick: Failed to process image #{source_file}: #{e.message}"
+                            next
                         end
 
                     end
